Use UTF-8 for Linux
The wchar_t in Linux is very wide, 32 bits, and thus almost unused,
which, in turn, results in poor support in tooling. And that's to be
expected when UTF-8 is the default. At the same time, Iconv, which is
part of glibc, is universally available. Thus we can use it to convert
UCS-2 into UTF-8 in Linux and always use UTF-8 in the library API.